+++ src/main/java/com/revature/RevRelay/services/UserService.java
@@ -1,6 +1,6 @@
 package com.revature.RevRelay.services;
 
-import com.revature.RevRelay.repository.UserRepository;
+import com.revature.RevRelay.repositories.UserRepository;
 import com.revature.RevRelay.models.User;
 import com.revature.RevRelay.models.dtos.UserRegisterAuthRequest;
 import com.revature.RevRelay.utils.JwtUtil;
